fmd: change cc owners attribute from uint8 to unit16

+++ drivers/net/ethernet/freescale/fman/Peripherals/FM/Pcd/fm_cc.h
@@ -328,7 +328,7 @@ typedef struct
     uint8_t             offset;
     uint8_t             prsArrayOffset;
     bool                ctrlFlow;
-    uint8_t             owners;
+    uint16_t            owners;
 
     uint8_t             ccKeySizeAccExtraction;
     uint8_t             sizeOfExtraction;
@@ -363,7 +363,7 @@ typedef struct
 {
     t_FmPcdCcNode       *p_FmPcdCcNode;
     bool                occupied;
-    uint8_t             owners;
+    uint16_t            owners;
     volatile bool       lock;
 } t_FmPcdCcNodeArray;
 
@@ -384,7 +384,7 @@ typedef struct
     t_List              fmPortsLst;
     t_FmPcdLock         *p_Lock;
     uint8_t             numOfEntries;
-    uint8_t             owners;
+    uint16_t            owners;
     t_Handle            h_FmPcdCcSavedManipParams;
     bool                modifiedState;
     uint32_t            requiredAction;
